Add typescript support for jest
