+++ package.json
@@ -35,8 +35,14 @@
     "request-promise": "^4.2.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.14.6",
+    "@babel/preset-env": "^7.14.7",
+    "@babel/preset-typescript": "^7.14.5",
+    "@types/jest": "^26.0.23",
     "@types/mongodb": "^3.6.18",
     "apollo-server-testing": "2.25.0",
+    "babel-jest": "^27.0.6",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "casual": "1.6.2",
     "jest": "27.0.4",
     "ts-node": "^10.0.0",
